requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-Cors==4.0.0
python-dotenv==1.0.0
mysql-connector-python==8.2.0
reportlab==4.0.8
Werkzeug==3.0.1
